Refactor tRPC types and improve project structure
- Split tRPC types into modular files under src/types/
- Create separate files for auth, status, and common tRPC configurations
- Update main trpc.ts to export from new modular structure
- Improve type organization and imports
