+++ src/types/api/index.ts
@@ -1,0 +1,5 @@
+// Exportar todos los tipos de la API
+export * from './auth';
+export * from './status';
+
+// Agregar más exportaciones a medida que se creen más archivos de tipos
