+++ src/types/api/status.ts
@@ -1,0 +1,5 @@
+// Tipos para el endpoint de estado del sistema
+
+export type PingResponse = {
+  status: string;
+};
