+++ src/types/trpc/common.ts
@@ -1,0 +1,4 @@
+import { initTRPC } from '@trpc/server';
+
+// Inicializar tRPC
+export const t = initTRPC.create();
