+++ src/types/trpc/status.ts
@@ -1,0 +1,8 @@
+import { t } from '@/types/trpc/common';
+
+// Definir el router de estado
+export const statusRouter = t.router({
+  ping: t.procedure.query(async () => {
+    return { status: 'ok' };
+  }),
+});
